Transform.task/requirements.txt
duckdb==0.8.0
boto3==1.26.153
python-dotenv==1.0.0
pandas==2.0.2